+++ src/util/internal-case.ts
@@ -1,4 +1,4 @@
-import { camelCase, lowerCase } from 'lodash-es';
+import { camelCase, lowerCase } from 'es-toolkit/compat';
 
 const internalCase = (value: string | undefined) => camelCase(lowerCase(value));