+++ src/util/pascal-case.ts
@@ -1,4 +1,4 @@
-import { camelCase, upperFirst } from 'lodash-es';
+import { camelCase, upperFirst } from 'es-toolkit/compat';
 
 const pascalCase = (value: string | undefined) => upperFirst(camelCase(value));